+++ src/components/MypageContent.tsx
@@ -1,0 +1,5 @@
+const MypageContent = () => {
+  return <div>MypageContent</div>;
+};
+
+export default MypageContent;
