+++ src/components/SignUpContent.tsx
@@ -1,0 +1,5 @@
+const SignUpContent = () => {
+  return <div>SignUpContent</div>;
+};
+
+export default SignUpContent;
